Co-locate error types with where they are used
This avoids unnecessary imports and makes it easier to cleanup errors
once they are no longer used.
