--- cnd/src/storage/db/errors.rs
@@ -1,37 +1,0 @@
-//! Defines errors for the database module.
-//!
-//! These are single-error structs in contrast to enums. We use `anyhow::Result`
-//! in almost all of our APIs. There is little value in defining enums that
-//! describe a range of errors if you have to downcast from an anyhow::Error
-//! anyway.
-
-use crate::LocalSwapId;
-use comit::{OrderId, Side};
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("no swap exists in the database for id {0}")]
-pub struct NoSwapExists(pub LocalSwapId);
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("no order exists in the database for id {0}")]
-pub struct NoOrderExists(pub OrderId);
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("no order found in the database for swap id {0}")]
-pub struct NoOrderForSwap(pub LocalSwapId);
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("no secret hash found in database for swap {0}")]
-pub struct NoSecretHash(pub LocalSwapId);
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("Order {0} is no longer open and can therefore not be cancelled")]
-pub struct NotOpen(pub OrderId);
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("Order {0} is not settling and can therefore not be closed")]
-pub struct NotSettling(pub OrderId);
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-#[error("both params are side {0}")]
-pub struct SameSide(pub Side);
